+++ package.json
@@ -18,6 +18,7 @@
     "@supabase/auth-helpers-react": "^0.4.2",
     "@supabase/supabase-js": "^2.31.0",
     "@tanstack/react-query": "^4.32.0",
+    "firebase": "^10.3.1",
     "js-cookie": "^3.0.5",
     "next": "13.4.16",
     "next-compose-plugins": "^2.2.1",
