--- src/components/Commons/Notifications/index.ts
@@ -1,1 +1,0 @@
-export { Notifications } from "./Notifications";
