+++ src/components/Commons/index.ts
@@ -10,7 +10,6 @@ export * from "./Inputs";
 export { Label } from "./Label";
 export * from "./Menu";
 export { Modal, useModal } from "./Modal";
-export { Notifications } from "./Notifications";
 export { PROFILE_DETAIL_MODAL, ProfileDetail } from "./ProfileDetail";
 export * from "./ProjectDetail";
 export { RatioBox } from "./RatioBox";
